refactor ui trace graph detail helpers
